build(deploy/terraform/modules): abstract the creation of the components needed to create the application by creating terraform modules
the application requires a set configuration in order to funtion. the creation of these terraform modules adheres strictly to the architecture design of the application. by creating these modules, the developer can create any number of copies of the application. as the infrasture code has been created via the modules, the developer would only need to create instances of the app instead of writing terraform code for each components. the modules are generic enough to allow customization of values, without compromising the architecture design