feat(main.tf): set up terraform and create an instance of the app infrastructure
because of the terraform modules written for each component, the abstraction of the application components allows the developer to focus solely on the values of the configuration as opposed to create infrastructure code. the abstraction allows the developer to provide a data structure that defines values to feed into the modules that will in turn create the application infrastructure. the abstraction also allows the developer to easily make copies of the application either in the same project or in different projects, by simply creating more of these data structures